requirements.txt
pyspark==3.5.2
requests==2.32.3
beautifulsoup4==4.12.3
google-generativeai==0.8.0
python-dotenv==1.0.1
delta-spark==3.2.0
deltalake==0.19.2
duckdb==1.1.0
polars==1.9.0
plotly==5.24.1
pandas==2.2.2
seaborn==0.13.2
